imp(ante): refactor AnteHandler
