+++ x/evm/keeper/utils.go
@@ -1,7 +1,6 @@
 package keeper
 
 import (
-	"math"
 	"math/big"
 
 	errorsmod "cosmossdk.io/errors"
@@ -17,21 +16,15 @@ import (
 )
 
 // DeductTxCostsFromUserBalance it calculates the tx costs and deducts the fees
-// returns (effectiveFee, priority, error)
 func (k Keeper) DeductTxCostsFromUserBalance(
 	ctx sdk.Context,
 	msgEthTx evmtypes.MsgEthereumTx,
 	txData evmtypes.TxData,
 	denom string,
+	baseFee *big.Int,
 	homestead, istanbul, london bool,
-) (fees sdk.Coins, priority int64, err error) {
+) (fees sdk.Coins, err error) {
 	isContractCreation := txData.GetTo() == nil
-
-	// fetch sender account from signature
-	signerAcc, err := authante.GetSignerAcc(ctx, k.accountKeeper, msgEthTx.GetFrom())
-	if err != nil {
-		return nil, 0, errorsmod.Wrapf(err, "account not found for sender %s", msgEthTx.From)
-	}
 
 	gasLimit := txData.GetGas()
 
@@ -42,7 +35,7 @@ func (k Keeper) DeductTxCostsFromUserBalance(
 
 	intrinsicGas, err := core.IntrinsicGas(txData.GetData(), accessList, isContractCreation, homestead, istanbul)
 	if err != nil {
-		return nil, 0, errorsmod.Wrapf(
+		return nil, errorsmod.Wrapf(
 			err,
 			"failed to retrieve intrinsic gas, contract creation = %t; homestead = %t, istanbul = %t",
 			isContractCreation, homestead, istanbul,
@@ -51,53 +44,43 @@ func (k Keeper) DeductTxCostsFromUserBalance(
 
 	// intrinsic gas verification during CheckTx
 	if ctx.IsCheckTx() && gasLimit < intrinsicGas {
-		return nil, 0, errorsmod.Wrapf(
+		return nil, errorsmod.Wrapf(
 			errortypes.ErrOutOfGas,
 			"gas limit too low: %d (gas limit) < %d (intrinsic gas)", gasLimit, intrinsicGas,
 		)
 	}
 
-	var feeAmt *big.Int
-
-	baseFee := k.getBaseFee(ctx, london)
 	if baseFee != nil && txData.GetGasFeeCap().Cmp(baseFee) < 0 {
-		return nil, 0, errorsmod.Wrapf(errortypes.ErrInsufficientFee,
+		return nil, errorsmod.Wrapf(errortypes.ErrInsufficientFee,
 			"the tx gasfeecap is lower than the tx baseFee: %s (gasfeecap), %s (basefee) ",
 			txData.GetGasFeeCap(),
 			baseFee)
 	}
 
-	feeAmt = txData.EffectiveFee(baseFee)
+	feeAmt := txData.EffectiveFee(baseFee)
 	if feeAmt.Sign() == 0 {
 		// zero fee, no need to deduct
-		return sdk.Coins{}, 0, nil
+		return sdk.Coins{}, nil
 	}
 
-	fees = sdk.Coins{sdk.NewCoin(denom, sdkmath.NewIntFromBigInt(feeAmt))}
+	fees = sdk.Coins{{Denom: denom, Amount: sdkmath.NewIntFromBigInt(feeAmt)}}
+
+	// fetch sender account from signature
+	signerAcc, err := authante.GetSignerAcc(ctx, k.accountKeeper, msgEthTx.GetFrom())
+	if err != nil {
+		return nil, errorsmod.Wrapf(err, "account not found for sender %s", msgEthTx.From)
+	}
 
 	// deduct the full gas cost from the user balance
 	if err := authante.DeductFees(k.bankKeeper, ctx, signerAcc, fees); err != nil {
-		return nil, 0, errorsmod.Wrapf(
+		return nil, errorsmod.Wrapf(
 			err,
 			"failed to deduct full gas cost %s from the user %s balance",
 			fees, msgEthTx.From,
 		)
 	}
 
-	// calculate priority based on effective gas price
-	tipPrice := txData.EffectiveGasPrice(baseFee)
-	// if london hardfork is not enabled, tipPrice is the gasPrice
-	if baseFee != nil {
-		tipPrice = new(big.Int).Sub(tipPrice, baseFee)
-	}
-	priorityBig := new(big.Int).Quo(tipPrice, evmtypes.DefaultPriorityReduction.BigInt())
-	if !priorityBig.IsInt64() {
-		priority = math.MaxInt64
-	} else {
-		priority = priorityBig.Int64()
-	}
-
-	return fees, priority, nil
+	return fees, nil
 }
 
 // CheckSenderBalance validates that the tx cost value is positive and that the
